+++ app/javascript/vue/components/Filter/Facets/CollectionObject/FacetRepository.vue
@@ -36,17 +36,23 @@ const emit = defineEmits(['update:modelValue'])
 
 const params = computed({
   get: () => props.modelValue,
-  set: value => emit('update:modelValue', value)
+  set: (value) => emit('update:modelValue', value)
 })
+
+const repositorySelected = ref()
 
 watch(
   () => props.modelValue.repository_id,
-  () => {
-    repositorySelected.value = undefined
+  (newId, oldVal) => {
+    if (oldVal && !newId) {
+      repositorySelected.value = undefined
+    }
   }
 )
 
-const repositorySelected = ref()
+watch(repositorySelected, (newVal) => {
+  params.value.repository_id = newVal?.id
+})
 
 onBeforeMount(() => {
   const urlParams = URLParamsToJSON(location.href)
@@ -54,25 +60,23 @@ onBeforeMount(() => {
   params.value.repository_id = urlParams.repository_id
 
   if (urlParams.repository_id) {
-    Repository.find(urlParams.repository_id).then(response => {
+    Repository.find(urlParams.repository_id).then((response) => {
       setRepository(response.body)
     })
   }
 })
 
-const setRepository = repository => {
+const setRepository = (repository) => {
   repositorySelected.value = repository
-  params.value.repository_id = repository.id
 }
 
 const unsetRepository = () => {
   repositorySelected.value = undefined
-  params.value.repository_id = undefined
 }
 </script>
 
 <style scoped>
-  :deep(.vue-autocomplete-input) {
-    width: 100%
-  }
+:deep(.vue-autocomplete-input) {
+  width: 100%;
+}
 </style>