+++ app/javascript/vue/components/radials/object/components/asserted_distributions/asserted_distributions_annotator.vue
@@ -127,18 +127,6 @@ export default {
       )
     },
 
-    existingArea() {
-      return this.list.find(
-        (item) =>
-          item.geographic_area_id ===
-            this.asserted_distribution.geographic_area_id &&
-          (item.is_absent === null ? false : item.is_absent) ===
-            (this.asserted_distribution.is_absent === null
-              ? false
-              : this.asserted_distribution.is_absent)
-      )
-    },
-
     shapes() {
       return this.list.map((item) => {
         const shape = item.geographic_area.shape
@@ -190,6 +178,12 @@ export default {
     },
 
     saveAssertedDistribution() {
+      const createdObject = this.list.find(
+        (item) =>
+          item.geographic_area.id ===
+            this.asserted_distribution.geographic_area_id &&
+          !!this.asserted_distribution.is_absent === !!item.is_absent
+      )
       const params = {
         asserted_distribution: {
           ...this.asserted_distribution,
@@ -198,9 +192,9 @@ export default {
         ...EXTEND_PARAMS
       }
 
-      const saveRequest = !this.existingArea
-        ? AssertedDistribution.create(params)
-        : AssertedDistribution.update(this.existingArea.id, params)
+      const saveRequest = createdObject
+        ? AssertedDistribution.update(createdObject.id, params)
+        : AssertedDistribution.create(params)
 
       saveRequest.then(({ body }) => {
         TW.workbench.alert.create(
@@ -212,7 +206,7 @@ export default {
     },
 
     removeCitation(item) {
-      const asserted_distribution = {
+      const payload = {
         citations_attributes: [
           {
             id: item.id,
@@ -222,7 +216,7 @@ export default {
       }
 
       AssertedDistribution.update(this.asserted_distribution.id, {
-        asserted_distribution,
+        asserted_distribution: payload,
         ...EXTEND_PARAMS
       }).then(({ body }) => {
         this.addToList(body)