+++ app/javascript/vue/tasks/otu/browse/components/biologicalAssociations/BiologicalAssociationRelated.vue
@@ -9,6 +9,7 @@
     </template>
 
     <template #body>
+      <VSpinner v-if="isLoading" />
       <ul>
         <li
           v-for="item in biologicalAssociations"
@@ -20,24 +21,70 @@
   </VModal>
   <VBtn
     color="primary"
-    :disabled="!biologicalAssociations.length"
     @click="isModalVisible = true"
   >
-    Related ({{ biologicalAssociations.length }})
+    Related ({{ total }})
   </VBtn>
 </template>
 
 <script setup>
-import { ref } from 'vue'
+import { BiologicalAssociation } from 'routes/endpoints'
+import { ref, onBeforeMount, watch } from 'vue'
+import { OTU, COLLECTION_OBJECT } from 'constants/index.js'
 import VModal from 'components/ui/Modal.vue'
 import VBtn from 'components/ui/VBtn/index.vue'
+import getPagination from 'helpers/getPagination'
+import VSpinner from 'components/spinner.vue'
 
-defineProps({
-  biologicalAssociations: {
-    type: Array,
-    default: () => []
+const param = {
+  [OTU]: 'otu_id',
+  [COLLECTION_OBJECT]: 'collection_object_id'
+}
+
+const props = defineProps({
+  current: {
+    type: Object,
+    required: true
+  },
+
+  itemId: {
+    type: Number,
+    required: true
+  },
+
+  itemType: {
+    type: String,
+    required: true
   }
 })
 
 const isModalVisible = ref(false)
+const biologicalAssociations = ref([])
+const total = ref('?')
+const isLoading = ref(false)
+
+onBeforeMount(async () => {
+  const payload = {
+    [param[props.itemType]]: [props.itemId],
+    per: 1
+  }
+  const response = await BiologicalAssociation.where(payload)
+
+  total.value = getPagination(response).total - 1
+})
+
+watch(isModalVisible, async (newVal) => {
+  if (newVal && !biologicalAssociations.value.length) {
+    isLoading.value = true
+    const { body } = await BiologicalAssociation.all({
+      [param[props.itemType]]: [props.itemId]
+    })
+
+    isLoading.value = false
+
+    biologicalAssociations.value = body.filter(
+      (item) => item.id !== props.current.id
+    )
+  }
+})
 </script>