+++ app/models/taxon_name_relationship.rb
@@ -379,7 +379,7 @@ class TaxonNameRelationship < ApplicationRecord
     end
 
     unless self.type_class.blank? # only validate if it is set
-      if object_taxon_name
+      if object_taxon_name && object_taxon_name.rank_class
         if object_taxon_name.type == 'Protonym' || object_taxon_name.type == 'Hybrid'
           unless self.type_class.valid_object_ranks.include?(self.object_taxon_name.rank_string)
             errors.add(:object_taxon_name_id, "#{self.object_taxon_name.rank_class.rank_name.capitalize} rank of #{self.object_taxon_name.cached_html} is not compatible with the #{self.subject_status} relationship")
@@ -388,7 +388,7 @@ class TaxonNameRelationship < ApplicationRecord
         end
       end
 
-      if subject_taxon_name
+      if subject_taxon_name && subject_taxon_name.rank_class
         if subject_taxon_name.type == 'Protonym' || subject_taxon_name.type == 'Hybrid'
           unless self.type_class.valid_subject_ranks.include?(self.subject_taxon_name.rank_string)
             soft_validations.add(:subject_taxon_name_id, "#{self.subject_taxon_name.rank_class.rank_name.capitalize} rank of #{self.subject_taxon_name.cached_html} is not compatible with the #{self.subject_status} relationship")