+++ lib/batch_load/import/taxon_names/nomen_interpreter.rb
@@ -21,9 +21,9 @@ module BatchLoad
 
     # @param [Hash] args
     def initialize(nomenclature_code: nil, parent_taxon_name_id: nil, also_create_otu: false, **args)
-      @nomenclature_code = nomenclature_code
-      @parent_taxon_name_id = parent_taxon_name_id
-      @also_create_otu = also_create_otu
+      @nomenclature_code = nomenclature_code.presence
+      @parent_taxon_name_id = parent_taxon_name_id.presence
+      @also_create_otu = also_create_otu.presence
 
       super(**args)
     end
@@ -42,7 +42,6 @@ module BatchLoad
       Project.find(@project_id).root_taxon_name
     end
 
-
     # @return [Integer]
     # delegate :id, to: :parent_taxon_name, prefix: true
 
@@ -53,10 +52,11 @@ module BatchLoad
       i = 0
       taxon_names = {}
 
-      csv.each do |row|
-        i += 1
+      csv.each_with_index do |row, i|
 
         parse_result = BatchLoad::RowParse.new
+        parse_result.row_number = i # check vs. header etc.
+
         parse_result.objects[:original_taxon_name] = []
         parse_result.objects[:taxon_name] = []
         parse_result.objects[:taxon_name_relationship] = []
@@ -67,45 +67,51 @@ module BatchLoad
         begin
           next if ['complex', 'species group', 'series', 'variety', 'unidentified'].include?(row['rank'])
 
+          rank = Ranks.lookup(@nomenclature_code.to_sym, row['rank'])
+
+          parse_result.parse_errors.push 'Unknown rank.' if rank.blank?
+
           protonym_attributes = {
             name: row['taxon_name'],
             year_of_publication: year_of_publication(row['author_year']),
-            rank_class: Ranks.lookup(@nomenclature_code.to_sym, row['rank']),
+            rank_class: rank,
             by: user,
             project:,
             verbatim_author: verbatim_author(row['author_year']),
-            taxon_name_authors_attributes: taxon_name_authors_attributes(verbatim_author(row['author_year']))
+
+            # People are not created at this point
+            #  taxon_name_authors_attributes: taxon_name_authors_attributes(verbatim_author(row['author_year']))
           }
 
-        # Not implemented
+          # Not implemented
 
-        # if row['original_name']
-        #   original_protonym_attributes = {
-        #     verbatim_name: row['original_name'],
-        #     name: row['original_name'].split(' ')[-1],
-        #     year_of_publication: year_of_publication(row['author_year']),
-        #     rank_class: Ranks.lookup(@nomenclature_code.to_sym, row['original_rank']),
-        #     parent: parent_taxon_name,
-        #     by: user,
-        #     project:,
-        #     verbatim_author: verbatim_author(row['author_year']),
-        #     taxon_name_authors_attributes: taxon_name_authors_attributes(verbatim_author(row['author_year']))
-        #   }
+          # if row['original_name']
+          #   original_protonym_attributes = {
+          #     verbatim_name: row['original_name'],
+          #     name: row['original_name'].split(' ')[-1],
+          #     year_of_publication: year_of_publication(row['author_year']),
+          #     rank_class: Ranks.lookup(@nomenclature_code.to_sym, row['original_rank']),
+          #     parent: parent_taxon_name,
+          #     by: user,
+          #     project:,
+          #     verbatim_author: verbatim_author(row['author_year']),
+          #     taxon_name_authors_attributes: taxon_name_authors_attributes(verbatim_author(row['author_year']))
+          #   }
 
-        #   original_protonym = Protonym.new(original_protonym_attributes)
+          #   original_protonym = Protonym.new(original_protonym_attributes)
 
-        #   if row['original_rank'] == 'genus'
-        #     protonym_attributes[:original_genus] = original_protonym
-        #   elsif row['original_rank'] == 'subgenus'
-        #     protonym_attributes[:original_subgenus] = original_protonym
-        #   elsif row['original_rank'] == 'species'
-        #     protonym_attributes[:original_species] = original_protonym
-        #   elsif row['original_rank'] == 'subspecies'
-        #     protonym_attributes[:original_subspecies] = original_protonym
-        #   end
+          #   if row['original_rank'] == 'genus'
+          #     protonym_attributes[:original_genus] = original_protonym
+          #   elsif row['original_rank'] == 'subgenus'
+          #     protonym_attributes[:original_subgenus] = original_protonym
+          #   elsif row['original_rank'] == 'species'
+          #     protonym_attributes[:original_species] = original_protonym
+          #   elsif row['original_rank'] == 'subspecies'
+          #     protonym_attributes[:original_subspecies] = original_protonym
+          #   end
 
-        #   parse_result.objects[:original_taxon_name].push original_protonym
-        # end
+          #   parse_result.objects[:original_taxon_name].push original_protonym
+          # end
 
           p = Protonym.new(protonym_attributes)
 
@@ -115,10 +121,14 @@ module BatchLoad
 
           taxon_names[taxon_name_id] = p
 
-          if taxon_names[parent_id].nil?
+          if parent_id.blank?
             p.parent_id = parent_taxon_name_id
           else
-            p.parent = taxon_names[parent_id]
+            if taxon_names[parent_id]
+              p.parent = taxon_names[parent_id]
+            else
+              parse_result.parse_errors.push 'Parent ID is not defined at this point! Row out of order?'
+            end
           end
 
           # TaxonNameRelationship
@@ -139,6 +149,7 @@ module BatchLoad
                 parse_result.objects[:taxon_name_relationship].push taxon_name_relationship
               end
             rescue NameError
+              parse_result.parse_errors.push 'Unknown taxon name relationship'
             end
           end
 
@@ -146,9 +157,12 @@ module BatchLoad
 
           # TODO: add to index, not here
           if name_nomen_classification = row['name_nomen_classification']
-            if c = name_nomen_classification.safe_constantize
-              # if EXCEPTED_FORM_TAXON_NAME_CLASSIFICATIONS.include?(name_nomen_classification)
-              p.taxon_name_classifications_attributes = [ {type: name_nomen_classification} ]
+            begin
+              if c = name_nomen_classification.safe_constantize
+                p.taxon_name_classifications_attributes = [ {type: name_nomen_classification} ]
+              end
+            rescue NameError
+              parse_result.parse_errors.push 'Unknown taxon name classification'
             end
           end
 
@@ -208,39 +222,39 @@ module BatchLoad
       author_year.gsub(/\,+\s*\d\d\d\d/, '')
     end
 
-    # @param [String] author_info
-    # @return [Array]
-    def taxon_name_authors_attributes(author_info)
-      return [] if author_info.blank?
-      multiple_author_query = 'and'
-      multiple_author_index = author_info.index(multiple_author_query)
-      split_author_info = multiple_author_index.nil? ? [author_info] : author_info.split(multiple_author_query)
-      author_infos = []
-
-      split_author_info.each do |author_str|
-        author_infos.push(author_info(author_str)) if author_str != 'NA' && author_str != 'unpublished'
-      end
-
-      author_infos
-    end
+    # # @param [String] author_info
+    # # @return [Array]
+    # def taxon_name_authors_attributes(author_info)
+    #   return [] if author_info.blank?
+    #   multiple_author_query = 'and'
+    #   multiple_author_index = author_info.index(multiple_author_query)
+    #   split_author_info = multiple_author_index.nil? ? [author_info] : author_info.split(multiple_author_query)
+    #   author_infos = []
+    #
+    #   split_author_info.each do |author_str|
+    #     author_infos.push(author_info(author_str)) if author_str != 'NA' && author_str != 'unpublished'
+    #   end
+    #
+    #   author_infos
+    # end
 
     # @param [String] author_string
     # @return [Hash]
-    def author_info(author_string)
-      seperator_query = ' '
-      separator_index = author_string.index(seperator_query)
-
-      last_name = author_string
-      first_name = ''
-
-      if !separator_index.nil?
-        separator_index += seperator_query.length
-        split_author_info = author_string.split(seperator_query)
-        last_name = split_author_info[0]
-        first_name = split_author_info[1]
-      end
-
-      { last_name:, first_name:, suffix: 'suffix' }
-    end
+    #  def author_info(author_string)
+    #  seperator_query = ' '
+    #  separator_index = author_string.index(seperator_query)
+    #
+    #  last_name = author_string
+    #  first_name = ''
+    #
+    #  if !separator_index.nil?
+    #    separator_index += seperator_query.length
+    #    split_author_info = author_string.split(seperator_query)
+    #    last_name = split_author_info[0]
+    #    first_name = split_author_info[1]
+    #  end
+    #
+    #  { last_name:, first_name: }
+    #end
   end
 end